+++ modules/php-nginx-mysql/jefe-cli.sh
@@ -3,16 +3,16 @@
 # php-nginx-mysql jefe-cli.sh
 #
 
-# Load utilities
-source ~/.jefe-cli/libs/utilities.sh
-
+if [[ -f  "$PROYECT_DIR/.env" ]]; then
+    load_dotenv
+fi
 # Docker compose var env configuration.
 docker_env() {
     puts "Docker compose var env configuration." BLUE
-    #     if [[ ! -f ".jefe/.env" ]]; then
-    #         cp .jefe/default.env .jefe/.env
+    #     if [[ ! -f "$PROYECT_DIR/.env" ]]; then
+    #         cp $PROYECT_DIR/default.env $PROYECT_DIR/.env
     #     fi
-    echo "" > .jefe/.env
+    echo "" > $PROYECT_DIR/.env
     set_dotenv PROJECT_TYPE $project_type
     puts "Write project name (default $project_type):" MAGENTA
     read proyect_name
@@ -64,6 +64,45 @@ docker_env() {
     else
         set_dotenv DB_PASSWORD $option
     fi
+    puts "Select framework:" MAGENTA
+    flag=true
+    while [ $flag = true ]; do
+        puts "1) None"
+        puts "2) Laravel"
+        puts "3) CakePHP"
+        puts "Type the option (number) that you want(digit), followed by [ENTER]:"
+        read option
+
+        case $option in
+            1)
+                framework=None
+                puts "Write DocumentRoot (default /var/www/html):" MAGENTA
+                read option
+                if [ -z $option ]; then
+                    document_root='/var/www/html'
+                else
+                    document_root=$option
+                fi
+                flag=false
+                ;;
+            2)
+                framework=Laravel
+                document_root='/var/www/html'
+                flag=false
+                ;;
+            3)
+                framework=CakePHP
+                document_root='/var/www/html/webroot'
+                flag=false
+                ;;
+            *)
+                puts "Wrong option" RED
+                flag=true
+                ;;
+        esac
+    done
+    set_dotenv FRAMEWORK $framework
+    set_dotenv DOCUMENT_ROOT $document_root
     puts "Database root password is password" YELLOW
     set_dotenv DB_ROOT_PASSWORD "password"
     puts "phpMyAdmin url: phpmyadmin.$vhost" YELLOW
@@ -71,13 +110,29 @@ docker_env() {
 
 # Add vhost of /etc/hosts file
 set_vhost(){
+    remove_vhost # Remove old vhost
     if [ ! "$( grep jefe-cli_wordpress /etc/hosts )" ]; then
         puts "Setting vhost..." BLUE
         load_dotenv
-        sudo sh -c "echo '127.0.0.1     $VHOST # ----- jefe-cli_$project_name' >> /etc/hosts"
-        sudo sh -c "echo '127.0.0.1     phpmyadmin.$VHOST # ----- jefe-cli_$project_name' >> /etc/hosts"
+        hosts="$( echo "$VHOST" | tr ',' ' ' )"
+        for host in $hosts; do
+            sudo sh -c "echo '127.0.0.1     $host # ----- jefe-cli_$project_name' >> /etc/hosts"
+            sudo sh -c "echo '127.0.0.1     phpmyadmin.$host # ----- jefe-cli_$project_name' >> /etc/hosts"
+        done
         puts "Done." GREEN
     fi
+}
+
+# Fix permisions of the proyect folder
+permissions(){
+    load_dotenv
+    puts "Setting permissions..." BLUE
+    cd $PROYECT_DIR
+        if id "www-data" >/dev/null 2>&1; then
+            sudo chown -R "$USER:www-data" $project_root
+        fi
+    cd ..
+    puts "Done." GREEN
 }
 
 # Create dump of the database of the proyect.
@@ -194,6 +249,7 @@ EOF
     fi
 }
 
+# Execute the command "composer install" in workdir folder
 composer_install() {
     e=$1
     if [ -z "${e}" ]; then
@@ -208,6 +264,7 @@ composer_install() {
     fi
 }
 
+# Execute the command "composer update" in workdir folder
 composer_update() {
     e=$1
     if [ -z "${e}" ]; then
@@ -221,3 +278,49 @@ composer_update() {
         ssh ${user}@${host} -p $port "cd ${public_dir}/; composer update"
     fi
 }
+
+if [[ $FRAMEWORK == "Laravel" ]]; then
+# Execute the command "php artisan migrate" in workdir folder. Running laravel migrations
+migrate() {
+        usage= cat <<EOF
+migrate [-e] [--environment] [-f] [--force] [--refresh] [--refresh-seed] [-h] [--help]
+
+Arguments:
+    -e, --environment		Set environment to import dump. Default is docker
+    -f, --force			Force Migrations to run in production (migrate
+        --refresh			Roll back all of your migrations and then execute the  migrate command
+        --refresh-seed			Roll back all of your migrations, execute the  migrate command and run all database seed
+    -h, --help			Print Help (this message) and exit
+EOF
+        # set an initial value for the flag
+        ENVIRONMENT="docker"
+        MIGRATE_OPTION=""
+
+        # read the options
+        OPTS=`getopt -o e:fh --long environment:,force,refresh,refresh-seed,help -n 'jefe' -- "$@"`
+        if [ $? != 0 ]; then puts "Invalid options." RED; exit 1; fi
+        eval set -- "$OPTS"
+
+        # extract options and their arguments into variables.
+        load_dotenv
+        while true ; do
+            case "$1" in
+                -e|--environment) ENVIRONMENT=$2 ; shift 2 ;;
+                -f|--force) MIGRATE_OPTION=' --force' ; shift 2 ;;
+                --refresh) MIGRATE_OPTION=':refresh' ; shift 2 ;;
+                --refresh-seed) MIGRATE_OPTION=':refresh --seed' ; shift 2 ;;
+                -h|--help) echo $usage ; exit 1 ; shift ;;
+                --) shift ; break ;;
+                *) echo "Internal error!" ; exit 1 ;;
+            esac
+        done
+
+        docker exec -it ${project_name}_php bash -c "php artisan migrate${MIGRATE_OPTION}"
+    }
+
+    # Execute the command "php artisan db:seed" in workdir folder. Run all laravel database seeds
+    seed() {
+        load_dotenv
+        docker exec -it ${project_name}_php bash -c 'php artisan db:seed'
+    }
+fi